requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
httpx==0.27.2
langgraph==0.2.14
python-dotenv==1.0.1
sqlalchemy==2.0.35
